Add files via upload
Improved run time for reference file generation. Changed to using up to an order 3 polynomial in beta for the estimation since it was found that for wider ranges of beta the second degree used prior was insufficient.
Some changes were done to the estimation function to make it simpler and easier to understand.